feat(euler): Fixes problem 3. Changes type in Math to long instead of int.

+++ src/main/java/euler/problem1/MultiplesOf3And5.java
@@ -8,8 +8,8 @@ import euler.util.MathUtil;
 public class MultiplesOf3And5 {
 
     public static void main(String[] args) {
-        List<Integer> range = MathUtil.range(0, 999);
-        List<Integer> numbers = MathUtil.getDivisibleByEither(range, Arrays.asList(3, 5));
+        List<Long> range = MathUtil.range(0, 999);
+        List<Long> numbers = MathUtil.getDivisibleByEither(range, Arrays.asList(3L, 5L));
         System.out.println("Sum: " + MathUtil.sum(numbers));
     }
 }
